Exercício 1.2 das aulas de DSD
Write a Java class that allows parallel search in an array of integer.
It provides the following static method: public static int
parallelSearch(int x, int[] A, int numThreads)
This method creates as many threads as specified by numThreads, divides
the array A into that many parts,  and gives each thread a part of the
array to search for sequentially.
If any thread finds x, then it returns an index i such that A[i] = x.
Otherwise, the method returns -1.